requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
distro==1.9.0
fastapi==0.115.14
filelock==3.18.0
fsspec==2025.5.1
groq==0.29.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.33.1
idna==3.10
Jinja2==3.1.6
jiter==0.10.0
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.5
numpy==2.3.1
openai==1.92.2
packaging==25.0
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
safetensors==0.5.3
sniffio==1.3.1
starlette==0.46.2
sympy==1.14.0
tokenizers==0.21.2
torch==2.7.1
tqdm==4.67.1
transformers==4.53.0
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.5.0
uvicorn==0.34.3
